requirements.txt
Django==5.2.7
python-decouple==3.8
requests==2.31.0
bleach==6.1.0
django-apscheduler==0.6.2
APScheduler==3.10.4
